Add Booking Controller
